% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot.R
\name{int_scatter}
\alias{int_scatter}
\title{A default test function}
\usage{
int_scatter(user_func)
}
\arguments{
\item{user_func}{A function which will be called several times. TODO: fill in the details once they're known.}
}
\value{
Nothing, used for side effects.
}
\description{
A default test function
}
