% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscatr.R
\name{iscatr-shiny}
\alias{iscatr-shiny}
\alias{iscatrOutput}
\alias{renderIscatr}
\title{Shiny bindings for iscatr}
\usage{
iscatrOutput(outputId, width = "100\%", height = "400px")

renderIscatr(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a iscatr}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using iscatr within Shiny
applications and interactive Rmd documents.
}
